==> PaDeMoS/3a_plotNBIResults.R <==
rm(list=ls())
setwd("/home/hanna/Dropbox/Paper_Hanna/datausw/NRIResults/")
load("LM_NRI_Bio_All.RData")
load("LM_NRI_Veg_All.RData")
load("LM_NRI_Bio_WV.RData")
load("LM_NRI_Veg_WV.RData")
load("LM_NRI_Bio_QB.RData")
load("LM_NRI_Veg_QB.RData")
load("LM_NRI_Bio_RE.RData")
load("LM_NRI_Veg_RE.RData")

pdf("/home/hanna/Dropbox/Paper_Hanna/abb/Nri_Vegcover.pdf")
par(mfrow=c(2,2),oma=c(0,0,0,0),mar=c(5,5,0.5,0.5))
plot(lm_nri_Veg_All, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Veg_WV, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Veg_QB, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Veg_RE, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
dev.off()

pdf("/home/hanna/Dropbox/Paper_Hanna/abb/Nri_Biomass.pdf")
par(mfrow=c(2,2),oma=c(0,0,0,0),mar=c(5,5,0.5,0.5))
plot(lm_nri_Bio_All, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Bio_WV, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Bio_QB, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
plot(lm_nri_Bio_RE, constraint="p.value<0.01",range=c(0,1),xlim=c(305,1000),ylim=c(305,1000))
dev.off()
